Add test_user_cannot_edit_control_from_tree_view
